Refactoring..
