+++ app/src/constants.js
@@ -1,0 +1,1 @@
+export const localServerUrl = 'http://localhost:8890/'
