+++ app/src/utils/queue/hasBack.js
@@ -1,0 +1,3 @@
+export default function hasBack () {
+  return musicCursor >= 2
+}
