+++ app/src/utils/queue/hasFront.js
@@ -1,0 +1,3 @@
+export default function hasFront () {
+  return musicCursor < musicQueue[currentKeyword].length
+}
