+++ app/src/utils/queue/queueFront.js
@@ -1,0 +1,8 @@
+export default function queueFront (query) {
+  if (hasFront()) {
+    console.log('cached Front')
+    showSearchResult(musicQueue[currentKeyword][musicCursor++])
+  } else {
+    searchQuery(query)
+  }
+}
