+++ app/src/utils/rpc/searchQuery.js
@@ -1,0 +1,5 @@
+export default function searchQuery (query) {
+  ipcRenderer.send('youtube-search-query', query)
+  musicQueue[currentKeyword].push({ query })
+  musicCursor = musicQueue[currentKeyword].length
+}
