+++ app/src/utils/rpc/searchbar.js
@@ -1,0 +1,12 @@
+import find from './find'
+
+export default function searchBar () {
+  // useRef TODO
+  const val = document.getElementById('Search-Form').value
+  document.getElementById('Search-Form').value = ''
+  console.log(val)
+  if (!val) {
+    return
+  }
+  find(val)
+}
