+++ package.json
@@ -57,6 +57,7 @@
     "dependencies": {
         "@babel/plugin-proposal-class-properties": "^7.12.1",
         "@babel/runtime": "^7.12.5",
+        "@reduxjs/toolkit": "^1.5.0",
         "babel-plugin-transform-async-to-generator": "^6.24.1",
         "bootstrap": "^4.5.3",
         "cfg": "0.0.2",
